#73 在开发过程中，不断的修改js,ui.xml,page文件，会导致不响应请求

+++ framework/org.tinygroup.fileresolver/src/main/java/org/tinygroup/fileresolver/impl/FileResolverImpl.java
@@ -351,7 +351,9 @@ public class FileResolverImpl implements FileResolver {
             if (!fileObject.isExist()) {
                 // 文件已经被删除
                 for (FileProcessor fileProcessor : fileProcessorList) {
-                    fileProcessor.delete(fileObject);
+                	if(fileProcessor.isMatch(fileObject)){//匹配后才能删除
+                		  fileProcessor.delete(fileObject);
+                	}
                 }
             } else {
                 tempMap.put(path, fileObject);
